requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pymongo==4.3.3
google-generativeai==0.3.2
python-multipart==0.0.6
aiohttp==3.8.6
httpx==0.26.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
aio-pika==9.3.0
motor==3.1.1
websockets==12.0
wsproto==1.2.0